Add GAN (experimental)
Adds an optional GAN training mode. In this mode, an adversarial discriminator network helps push the model to generate images that look similar to the training data. This increases GPU memory and compute use, but may give better results.

Currently there are no pretrained discriminator networks available, so the discriminator network is initialized randomly. It may take a few thousand samples for the discriminator to start having an effect.